Inicial commit
this version contains two word lists, a code to add the words from one word list to the official wordlist. yet this does not yet include a testing file nor a main python file, the only purpose of this is to complete the word list with more words without doing it manually.